lsvpd: Determine valid scsi page codes before issuing inquiries
Currently, based on the manufacturer name, subtypeDS and model name,
a matching template is found in scsi_templates.conf file for the device
discovered values. From the page codes retrieved from the scsi template
identified, sg_ll_inquiry() calls are made for those page codes
respectively. This is generating bunch of illegal requests to drives
that are unnecessary.

To fix this, read inquiry page with page code 0 and evpd 1 to get the
supported page codes, and issue inquiries only if the page code from
the template is present in the list of page codes supported.

Signed-off-by: Sathvika Vasireddy <sv@linux.ibm.com>